# docs/swagger.yaml
basePath: /
definitions:
...
  ftgo-finpro_internal_customerHandler_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  ftgo-finpro_internal_customerHandler_models.LoginResponse:
    properties:
      email:
        type: string
      name:
        type: string
      token:
        type: string
      wallet_balance:
        type: number
    type: object
  ftgo-finpro_internal_customerHandler_models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: Obie.kal22@gmail.com
    url: www.linkedin.com/in/obie-ananda-a87a64212
  description: API documentation for the FTGO PlashCash project.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: FTGO PlasCash Project
  version: "1.0"
paths:
...
  /login-customer:
    post:
      consumes:
      - application/json
      description: This endpoint authenticates a customer using their email and password
        and returns a JWT token.
      parameters:
      - description: Request body for logging in a customer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ftgo-finpro_internal_customerHandler_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token and customer details
          schema:
            $ref: '#/definitions/ftgo-finpro_internal_customerHandler_models.LoginResponse'
      summary: Login a customer
      tags:
      - Customers
...
  /register-customer:
    post:
      consumes:
      - application/json
      description: This endpoint registers a new customer with name, email, and password.
      parameters:
      - description: Request body for registering a customer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ftgo-finpro_internal_customerHandler_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Customer registered successfully
          schema:
            additionalProperties: true
            type: object
      summary: Register a new customer
      tags:
      - Customers
...
swagger: "2.0"